scraper/requirements.txt
requests
beautifulsoup4
lxml
python-dateutil
ruamel.yaml
